+++ src/settings/BatchGroupConfigurationSettings/BatchVoucherExportsList.js
@@ -2,22 +2,29 @@ import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
-import { MultiColumnList } from '@folio/stripes/components';
+import {
+  MultiColumnList,
+  NoValue,
+} from '@folio/stripes/components';
 import {
   FolioFormattedTime,
   acqRowFormatter,
 } from '@folio/stripes-acq-components';
 
-import { BATCH_VOUCHER_EXPORT_STATUS_LABEL } from '../../common/constants';
+import {
+  BATCH_VOUCHER_EXPORT_STATUS,
+  BATCH_VOUCHER_EXPORT_STATUS_LABEL,
+} from '../../common/constants';
 import { RESULT_COUNT_INCREMENT } from './constants';
 import ExportVoucherButton from './ExportVoucherButton';
 
 const columnMapping = {
   date: <FormattedMessage id="ui-invoice.settings.BatchVoucherExports.date" />,
   status: <FormattedMessage id="ui-invoice.settings.BatchVoucherExports.status" />,
+  message: <FormattedMessage id="ui-invoice.settings.BatchVoucherExports.message" />,
   exportButton: ' ',
 };
-const visibleColumns = ['date', 'status', 'exportButton'];
+const visibleColumns = ['date', 'status', 'message', 'exportButton'];
 const rowProps = { alignLastColToEnd: true };
 
 export function BatchVoucherExportsList({ batchVoucherExports, format, onNeedMoreData, recordsCount }) {
@@ -35,6 +42,12 @@ export function BatchVoucherExportsList({ batchVoucherExports, format, onNeedMor
         />
       ),
       status: ({ status }) => BATCH_VOUCHER_EXPORT_STATUS_LABEL[status],
+      // eslint-disable-next-line react/prop-types
+      message: ({ message, status }) => (
+        status === BATCH_VOUCHER_EXPORT_STATUS.uploaded
+          ? <FormattedMessage id="ui-invoice.settings.BatchVoucherExports.message.uploaded" />
+          : message || <NoValue />
+      ),
     }),
     [format],
   );